[spaces] moved finitevolume -> fv
